test: reconstruct tests
