--- tests/integration/apps/async_http_app.py
@@ -1,9 +1,0 @@
-import asyncio
-
-from lcserve import serving
-
-
-@serving
-async def async_http(interval: int) -> str:
-    await asyncio.sleep(interval)
-    return 'Hello, world!'
