--- tests/integration/apps/sync_http_app.py
@@ -1,9 +1,0 @@
-import time
-
-from lcserve import serving
-
-
-@serving
-def sync_http(interval: int) -> str:
-    time.sleep(interval)
-    return 'Hello, world!'
